pylint & fix display_packet_crash
