+++ tests/fuzzer/display_packet_crash.py
@@ -7,7 +7,7 @@ import sys
 from ryu.controller import dpset
 from faucet import faucet
 from faucet import faucet_experimental_api
-import Fake
+import fake_packet
 
 
 def main():
@@ -25,8 +25,10 @@ def main():
     # make sure dps are running
     if application.valves_manager is not None:
         for valve in list(application.valves_manager.valves.values()):
+            state = valve.dp.dyn_finalized
             valve.dp.dyn_finalized = False
             valve.dp.running = True
+            valve.dp.dyn_finalized = state
 
     # create data from read file
     byte_data = None
@@ -37,9 +39,9 @@ def main():
 
     if byte_data is not None:
         # create fake packet
-        dp = Fake.Datapath(1)
-        msg = Fake.Message(datapath=dp, cookie=1524372928, port=1, data=byte_data, in_port=1)
-        pkt = Fake.RyuEvent(msg)
+        _dp = fake_packet.Datapath(1)
+        msg = fake_packet.Message(datapath=_dp, cookie=15243729, port=1, data=byte_data, in_port=1)
+        pkt = fake_packet.RyuEvent(msg)
 
         # send packet to faucet and display error produced
         application.packet_in_handler(pkt)
