+++ tests/fuzzer/fuzz_packet.py
@@ -8,7 +8,7 @@ from ryu.controller import dpset
 from faucet import faucet
 from faucet import faucet_experimental_api
 import afl
-import Fake
+import fake_packet
 
 ROUNDS = 1
 logging.disable(logging.CRITICAL)
@@ -24,8 +24,10 @@ def main():
     # make sure dps are running
     if application.valves_manager is not None:
         for valve in list(application.valves_manager.valves.values()):
+            state = valve.dp.dyn_finalized
             valve.dp.dyn_finalized = False
             valve.dp.running = True
+            valve.dp.dyn_finalized = state
 
     while afl.loop(ROUNDS):
         # receive input from afl
@@ -37,9 +39,9 @@ def main():
             continue
 
         # create fake packet
-        dp = Fake.Datapath(1)
-        msg = Fake.Message(datapath=dp, cookie=1524372928, port=1, data=data, in_port=1)
-        pkt = Fake.RyuEvent(msg)
+        _dp = fake_packet.Datapath(1)
+        msg = fake_packet.Message(datapath=_dp, cookie=15243729, port=1, data=data, in_port=1)
+        pkt = fake_packet.RyuEvent(msg)
 
         # send fake packet to faucet
         application.packet_in_handler(pkt)
